refactor(admin): AdminEvaluationResults sous 300 lignes — éclatement complet (H3)
Éclate AdminEvaluationResults.vue (836 l) en orchestrateur (181 l) +
composable + 3 sous-composants présentationnels, sur le modèle G1
(AdminUsers). Comportement et rendu identiques : mêmes appels services
(api.get '/evaluations', '/proxy/classes', '/proxy/matieres'), mêmes
routes (/admin/evaluations/:id/details, /(coordinateur|teacher)/
evaluations/:id/preview), mêmes classes CSS, mêmes textes/formats.

- useAdminEvaluationResults.js : toute la logique <script> (state,
  filtres, computed filteredEvaluations/stats, loadData, navigation,
  helpers statut/date) déplacée verbatim.
- EvalResultsStatsCards.vue : les 4 cartes de stats (CSS .stats-grid/
  .stat-card/... verbatim).
- EvalResultsFilters.vue : filtres enseignant/classe/matière/statut en
  v-model (defineModel) + emit reset (CSS .filters-* verbatim).
- EvalResultsList.vue : liste des évaluations + état vide (CSS
  .evaluation-*/.meta-*/.status-*/.btn-*/.empty-* verbatim) ; helpers
  d'affichage injectés en props depuis le composable (source unique).

Préfixe distinct EvalResults* + useAdminEvaluationResults pour éviter la
collision avec les Evaluation* d'un refactor frère.

Dette pré-existante conservée à l'identique (NON corrigée) :
- console.log de debug dans le filtre matière (filteredEvaluations).
- updateFilter() du script d'origine devient inutile (les filtres
  passent par v-model) : non réexporté vers la vue, mais conservé dans
  le composable à l'identique pour ne rien retirer du comportement.
- CSS .loading-state conservé dans la vue alors qu'aucun markup
  .loading-state n'existe (le loader est <ContentLoader>) — dead CSS
  déjà présent à l'origine, gardé verbatim.
- Sélecteurs combinés d'origine .empty-state,.error-state et
  .empty-icon,.error-icon scindés fidèlement : règles empty-* déplacées
  dans EvalResultsList (avec leur markup), règles error-* gardées dans
  la vue (avec leur markup) — corps de règles identiques.

Tests : useAdminEvaluationResults.test.js (9) + EvalResultsList.test.js
(4) = 13/13 verts. Build vite exit 0.
